perf: code-split LayerRendererPublic per layer type
Lazy-load AnimationInitializer, PaginatedCollection, LoadMoreCollection,
FilterableCollection, and LocaleSelector via next/dynamic so GSAP, Swiper,
and Zustand don't ship in the entry chunk for pages that don't use them.

Extract filter-layer DOM behavior + useFilterStore into FilterLayerBehavior
so the runtime only loads on pages that render a filter layer.

Cuts ~335 KB (22%) of eager JS on a slider+animation home page; larger
wins on simpler marketing pages where GSAP and Swiper drop out entirely.
